+++ crates/compiler/src/mml/tokenizer.rs
@@ -43,6 +43,8 @@ pub enum Token<'a> {
     Rest,
     Wait,
     PlayPitch,
+    PlayNoise,
+    DisableNoise,
     SetOctave,
     IncrementOctave,
     DecrementOctave,
@@ -248,7 +250,7 @@ fn is_unknown_u8(c: u8) -> bool {
         b'!' | b'@' | b'+' | b'-' | b'[' | b':' | b']' | b'^' | b'&' | b'C' | b's' | b'n'
         | b'l' | b'r' | b'w' | b'o' | b'>' | b'<' | b'v' | b'V' | b'p' | b'Q' | b'q' | b'~'
         | b'A' | b'G' | b'E' | b't' | b'T' | b'L' | b'%' | b'.' | b',' | b'|' | b'_' | b'{'
-        | b'}' | b'B' | b'D' | b'F' | b'I' | b'M' | b'P' => false,
+        | b'}' | b'B' | b'D' | b'F' | b'I' | b'M' | b'P' | b'N' => false,
         b'\\' => false,
         c if c.is_ascii_whitespace() => false,
         _ => true,
@@ -405,6 +407,15 @@ fn next_token<'a>(scanner: &mut Scanner<'a>) -> Option<TokenWithPosition<'a>> {
         b'I' => one_ascii_token!(Token::GainModeI),
 
         b'P' => one_ascii_token!(Token::PlayPitch),
+
+        b'N' => {
+            let c2 = scanner.second_byte();
+
+            match c2 {
+                Some(b'-') => two_ascii_token!(Token::DisableNoise),
+                _ => one_ascii_token!(Token::PlayNoise),
+            }
+        }
 
         // Pan might be 1 or 2 characters
         b'p' => {