+++ docs/changelog.md
@@ -10,6 +10,8 @@ Version 0.0.15
 
 Audio driver changes:
  * Added `play_pitch` instruction (`P` MML command)
+ * Added `play_noise` instruction (`N` MML command)
+ * Added `disable_noise` instruction (`N-` MML command)